Use Typed Arrays in websock.js
This commit converts websock.js to use Typed Arrays (and tweaks
rfb.js to ensure that it continues to function, since only Firefox
implements `%TypedArray%.prototype.slice`).  This provides two
advantages: first, we allocate a buffer ahead of time, meaning
the browser doesn't have to do any work dynamically increasing
the receive queue size.  Secondly, we are now able to pass around
Typed Array Views (e.g. `Uint8Array`), which are lightweight, and
don't involve copying.

The downside is that we initially allocate more memory -- we currently
start out with 4 MiB, and then automatically double when it looks like
the amount unused is getting to small.

The commit also explicitly adds a check to the compacting logic that
avoids calling the copy functions if `_rQlen === _rQi`.